+++ src/pydvl/value/shapley/actor.py
@@ -16,7 +16,7 @@ from pydvl.utils import Utility, get_running_avg_variance, maybe_progress
 from pydvl.utils.config import ParallelConfig
 from pydvl.utils.parallel.actor import Coordinator, RayActorWrapper, Worker
 from pydvl.utils.parallel.backend import init_parallel_backend
-from pydvl.value.valuationresult import ValuationStatus
+from pydvl.value import ValuationStatus
 
 if TYPE_CHECKING:
     from numpy.typing import NDArray
